+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # changelog for `handler_map`
 
+## Pending
+### Changes
+
+- `HandlerMap` now has a lifetime parameter, and is not forced to `'static`
+
 ## `0.1.0` - 2018-09-19
 
 Initial version!